Amazon Keyspaces-related database error handling (#86)

+++ test/tf/daemon.sh.tpl
@@ -78,21 +78,23 @@ aws s3 ls ${s3_log_url}/
 SEND_LOGS_FILE=/home/${ssh_user}/sendlogs.sh
 sudo tee $SEND_LOGS_FILE <<EOF
 #!/bin/bash
-# Send SIGHUP to the running binary, it will rotate the log using Lumberjack
-ps axf | grep capidaemon | grep -v grep | awk '{print "kill -s 1 " \$1}' | sh
-for f in /var/log/capillaries/*.gz;do
-  if [ -f \$f ]; then
-    # Lumberjack produces: capidaemon-2025-05-03T21-37-01.283.log.gz
-    # Add hostname to it: capidaemon-2025-05-03T21-37-01.283.ip-10-5-0-101.log.gz
-    fname=\$(basename -- "\$f")
-    fnamedatetime=\$(echo \$fname|cut -d'.' -f1)
-    fnamemillis=\$(echo \$fname|cut -d'.' -f2)
-    newfilepath=/var/log/capillaries/\$fnamedatetime.\$fnamemillis.\$HOSTNAME.log.gz
-    mv \$f \$newfilepath
-    aws s3 cp \$newfilepath ${s3_log_url}/
-    rm \$newfilepath
-  fi
-done
+if [ -s /var/log/capillaries/capidaemon.log ]; then
+  # Send SIGHUP to the running binary, it will rotate the log using Lumberjack
+  ps axf | grep capidaemon | grep -v grep | awk '{print "kill -s 1 " \$1}' | sh
+  for f in /var/log/capillaries/*.gz;do
+    if [ -e \$f ]; then
+      # Lumberjack produces: capidaemon-2025-05-03T21-37-01.283.log.gz
+      # Add hostname to it: capidaemon-2025-05-03T21-37-01.283.ip-10-5-0-101.log.gz
+      fname=\$(basename -- "\$f")
+      fnamedatetime=\$(echo \$fname|cut -d'.' -f1)
+      fnamemillis=\$(echo \$fname|cut -d'.' -f2)
+      newfilepath=/var/log/capillaries/\$fnamedatetime.\$fnamemillis.\$HOSTNAME.log.gz
+      mv \$f \$newfilepath
+      aws s3 cp \$newfilepath ${s3_log_url}/
+      rm \$newfilepath
+    fi
+  done
+fi
 EOF
 sudo chmod 744 $SEND_LOGS_FILE
 sudo su ${ssh_user} -c "echo \"*/5 * * * * $SEND_LOGS_FILE\" | crontab -"
@@ -105,4 +107,4 @@ sudo chown -R ${ssh_user} /home/${ssh_user}
 # Run daemon
 
 # AWS region is required because S3 bucket pointer is a URI, not a URL
-sudo su ${ssh_user} -c 'AWS_DEFAULT_REGION=${awsregion} CAPI_CASSANDRA_HOSTS="${cassandra_hosts}" CAPI_CASSANDRA_PORT=${cassandra_port} CAPI_CASSANDRA_USERNAME="${cassandra_username}" CAPI_CASSANDRA_PASSWORD="${cassandra_password}" CAPI_CASSANDRA_CA_PATH="/home/${ssh_user}/.ssh/sf-class2-root.crt" CAPI_CASSANDRA_ENABLE_HOST_VERIFICATION=false CAPI_CASSANDRA_KEYSPACE_REPLICATION_CONFIG="'"{ 'class' : 'SingleRegionStrategy'}"'" CAPI_CASSANDRA_CONSISTENCY=LOCAL_QUORUM CAPI_CASSANDRA_WRITER_WORKERS=20 CAPI_CASSANDRA_MIN_INSERTER_RATE=5 CAPI_THREAD_POOL_SIZE=8 CAPI_AMQP091_URL="${rabbitmq_url}" CAPI_CASSANDRA_TIMEOUT=15000 CAPI_PYCALC_INTERPRETER_PATH=python3 CAPI_LOG_LEVEL=info CAPI_LOG_FILE="/var/log/capillaries/capidaemon.log" /home/${ssh_user}/bin/capidaemon &>/dev/null &'
+sudo su ${ssh_user} -c 'AWS_DEFAULT_REGION=${awsregion} CAPI_CASSANDRA_HOSTS="${cassandra_hosts}" CAPI_CASSANDRA_PORT=${cassandra_port} CAPI_CASSANDRA_USERNAME="${cassandra_username}" CAPI_CASSANDRA_PASSWORD="${cassandra_password}" CAPI_CASSANDRA_CA_PATH="/home/${ssh_user}/.ssh/sf-class2-root.crt" CAPI_CASSANDRA_ENABLE_HOST_VERIFICATION=false CAPI_CASSANDRA_KEYSPACE_REPLICATION_CONFIG="'"{ 'class' : 'SingleRegionStrategy'}"'" CAPI_CASSANDRA_CONSISTENCY=LOCAL_QUORUM CAPI_CASSANDRA_WRITER_WORKERS=${writer_workers} CAPI_CASSANDRA_MIN_INSERTER_RATE=5 CAPI_THREAD_POOL_SIZE=${thread_pool_size} CAPI_AMQP091_URL="${rabbitmq_url}" CAPI_CASSANDRA_TIMEOUT=15000 CAPI_PYCALC_INTERPRETER_PATH=python3 CAPI_LOG_LEVEL=info CAPI_LOG_FILE="/var/log/capillaries/capidaemon.log" /home/${ssh_user}/bin/capidaemon &>/dev/null &'
